Refactor error and rescue more http error code
