+++ lib/woocommerce_api.rb
@@ -9,6 +9,7 @@ require "woocommerce_api/concerns/associations"
 require "woocommerce_api/concerns/attribute_assignment"
 require "woocommerce_api/concerns/singleton"
 require "woocommerce_api/client"
+require "woocommerce_api/client_error"
 require "woocommerce_api/oauth_client"
 require "woocommerce_api/resource"
 require "woocommerce_api/version"
